# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: GodsApp — Professional Security Auditing & Research Suite API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workspaces:
    get:
      operationId: listWorkspaces
      tags: [workspaces]
      summary: List all workspaces
      responses:
        "200":
          description: Workspace list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Workspace"

    post:
      operationId: createWorkspace
      tags: [workspaces]
      summary: Create a new workspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkspaceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
  /workspaces/{id}:
    get:
      operationId: getWorkspace
      tags: [workspaces]
      summary: Get workspace details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Workspace
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"

    patch:
      operationId: updateWorkspace
      tags: [workspaces]
      summary: Update workspace
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkspaceUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"

    delete:
      operationId: deleteWorkspace
      tags: [workspaces]
      summary: Delete a workspace
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /workspaces/{id}/stats:
    get:
      operationId: getWorkspaceStats
      tags: [workspaces]
      summary: Get workspace statistics
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Workspace stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceStats"

  # ── Scans ─────────────────────────────────────────────────────────────────
...
  /tools/network/dns:
    post:
      operationId: runDnsLookup
      tags: [tools]
      summary: Run DNS lookup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DnsLookupInput"
      responses:
        "200":
          description: DNS results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DnsLookupResult"
  /tools/network/whois:
    post:
      operationId: runWhoisLookup
      tags: [tools]
      summary: Run WHOIS lookup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WhoisInput"
      responses:
        "200":
          description: WHOIS result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WhoisResult"
  /tools/crypto/hash:
    post:
      operationId: calculateHash
      tags: [tools]
      summary: Calculate hash of text input
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HashInput"
      responses:
        "200":
          description: Hash results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HashResult"
  /tools/crypto/encode:
    post:
      operationId: encodeText
      tags: [tools]
      summary: Encode/decode text
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EncodeInput"
      responses:
        "200":
          description: Encoded/decoded result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EncodeResult"
  /tools/crypto/identify-hash:
    post:
      operationId: identifyHash
      tags: [tools]
      summary: Identify hash type from value
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IdentifyHashInput"
      responses:
        "200":
          description: Hash type candidates
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdentifyHashResult"
  /tools/web/ssl:
    post:
      operationId: analyzeSsl
      tags: [tools]
      summary: Analyze SSL/TLS configuration of a host
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SslAnalysisInput"
      responses:
        "200":
          description: SSL analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SslAnalysisResult"
  /tools/web/headers:
    post:
      operationId: analyzeHeaders
      tags: [tools]
      summary: Analyze HTTP security headers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HeaderAnalysisInput"
      responses:
        "200":
          description: Header analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HeaderAnalysisResult"
  /tools/web/jwt:
    post:
      operationId: analyzeJwt
      tags: [tools]
      summary: Decode and analyze a JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JwtAnalysisInput"
      responses:
        "200":
          description: JWT analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JwtAnalysisResult"
  /tools/exploitation/reverse-shell:
    post:
      operationId: generateReverseShell
      tags: [tools]
      summary: Generate reverse shell payload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReverseShellInput"
      responses:
        "200":
          description: Generated reverse shells
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReverseShellResult"
  /tools/intel/ip:
    post:
      operationId: lookupIpIntel
      tags: [tools]
      summary: Perform IP intelligence lookup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IpIntelInput"
      responses:
        "200":
          description: IP intelligence data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IpIntelResult"
  /tools/password/hibp:
    post:
      operationId: checkHibp
      tags: [tools]
      summary: Check email or password against HaveIBeenPwned
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HibpInput"
      responses:
        "200":
          description: HIBP check result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HibpResult"

  # ── Plugins ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: ["string", "null"]

    # ── Auth ────────────────────────────────────────────────────────────────
...
    Workspace:
      type: object
      required: [id, name, status, createdAt, updatedAt, color, findingCounts]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        client:
          type: ["string", "null"]
        scope:
          type: array
          items:
            type: string
        outOfScope:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [active, archived]
        color:
          type: string
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
        updatedAt:
          type: string
        lastActivity:
          type: ["string", "null"]
        findingCounts:
          $ref: "#/components/schemas/FindingCounts"
    WorkspaceInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        client:
          type: string
        scope:
          type: array
          items:
            type: string
        outOfScope:
          type: array
          items:
            type: string
        color:
          type: string
        tags:
          type: array
          items:
            type: string
    WorkspaceUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        client:
          type: ["string", "null"]
        scope:
          type: array
          items:
            type: string
        outOfScope:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [active, archived]
        color:
          type: string
        tags:
          type: array
          items:
            type: string
    WorkspaceStats:
      type: object
      required: [scanCount, findingCount, evidenceCount, findingsByDay]
      properties:
        scanCount:
          type: integer
        findingCount:
          type: integer
        evidenceCount:
          type: integer
        reportCount:
          type: integer
        findingCounts:
          $ref: "#/components/schemas/FindingCounts"
        findingsByDay:
          type: array
          items:
            $ref: "#/components/schemas/DailyCount"
        topTools:
          type: array
          items:
            $ref: "#/components/schemas/ToolCount"
    FindingCounts:
      type: object
      required: [critical, high, medium, low, info, total]
      properties:
        critical:
          type: integer
        high:
          type: integer
        medium:
          type: integer
        low:
          type: integer
        info:
          type: integer
        total:
          type: integer
    DailyCount:
      type: object
      required: [date, count]
      properties:
        date:
          type: string
        count:
          type: integer
    ToolCount:
      type: object
      required: [tool, count]
      properties:
        tool:
          type: string
        count:
          type: integer

    # ── Scans ───────────────────────────────────────────────────────────────
...
    DnsLookupInput:
      type: object
      required: [target]
      properties:
        target:
          type: string
        recordTypes:
          type: array
          items:
            type: string
        resolver:
          type: ["string", "null"]
    DnsLookupResult:
      type: object
      required: [target, records]
      properties:
        target:
          type: string
        records:
          type: array
          items:
            $ref: "#/components/schemas/DnsRecord"
    DnsRecord:
      type: object
      required: [type, value, ttl]
      properties:
        type:
          type: string
        value:
          type: string
        ttl:
          type: integer
    WhoisInput:
      type: object
      required: [target]
      properties:
        target:
          type: string
    WhoisResult:
      type: object
      required: [target, raw]
      properties:
        target:
          type: string
        registrar:
          type: ["string", "null"]
        registeredDate:
          type: ["string", "null"]
        expiresDate:
          type: ["string", "null"]
        nameservers:
          type: array
          items:
            type: string
        registrant:
          type: ["string", "null"]
        raw:
          type: string
    HashInput:
      type: object
      required: [text]
      properties:
        text:
          type: string
        algorithms:
          type: array
          items:
            type: string
    HashResult:
      type: object
      required: [input, hashes]
      properties:
        input:
          type: string
        hashes:
          type: object
          additionalProperties:
            type: string
    EncodeInput:
      type: object
      required: [text, encoding, mode]
      properties:
        text:
          type: string
        encoding:
          type: string
        mode:
          type: string
          enum: [encode, decode]
    EncodeResult:
      type: object
      required: [input, output, encoding, mode]
      properties:
        input:
          type: string
        output:
          type: string
        encoding:
          type: string
        mode:
          type: string
    IdentifyHashInput:
      type: object
      required: [hash]
      properties:
        hash:
          type: string
    IdentifyHashResult:
      type: object
      required: [hash, candidates]
      properties:
        hash:
          type: string
        candidates:
          type: array
          items:
            $ref: "#/components/schemas/HashCandidate"
    HashCandidate:
      type: object
      required: [algorithm, confidence, hashcatMode, johnFormat]
      properties:
        algorithm:
          type: string
        confidence:
          type: string
        hashcatMode:
          type: ["string", "null"]
        johnFormat:
          type: ["string", "null"]
    SslAnalysisInput:
      type: object
      required: [host]
      properties:
        host:
          type: string
        port:
          type: integer
    SslAnalysisResult:
      type: object
      required: [host, port, grade]
      properties:
        host:
          type: string
        port:
          type: integer
        grade:
          type: string
        subject:
          type: ["string", "null"]
        issuer:
          type: ["string", "null"]
        validFrom:
          type: ["string", "null"]
        validTo:
          type: ["string", "null"]
        daysUntilExpiry:
          type: ["integer", "null"]
        protocols:
          type: array
          items:
            type: string
        ciphers:
          type: array
          items:
            type: string
        vulnerabilities:
          type: array
          items:
            $ref: "#/components/schemas/SslVulnerability"
        securityHeaders:
          type: object
          additionalProperties:
            type: string
    SslVulnerability:
      type: object
      required: [name, vulnerable]
      properties:
        name:
          type: string
        vulnerable:
          type: boolean
        details:
          type: ["string", "null"]
    HeaderAnalysisInput:
      type: object
      required: [url]
      properties:
        url:
          type: string
    HeaderAnalysisResult:
      type: object
      required: [url, grade, headers, missing, issues]
      properties:
        url:
          type: string
        grade:
          type: string
        headers:
          type: object
          additionalProperties:
            type: string
        missing:
          type: array
          items:
            type: string
        issues:
          type: array
          items:
            $ref: "#/components/schemas/HeaderIssue"
    HeaderIssue:
      type: object
      required: [header, severity, description]
      properties:
        header:
          type: string
        severity:
          type: string
        description:
          type: string
        recommendation:
          type: ["string", "null"]
    JwtAnalysisInput:
      type: object
      required: [token]
      properties:
        token:
          type: string
        secret:
          type: ["string", "null"]
    JwtAnalysisResult:
      type: object
      required: [token, header, payload, issues]
      properties:
        token:
          type: string
        header:
          type: object
          additionalProperties: true
        payload:
          type: object
          additionalProperties: true
        algorithm:
          type: ["string", "null"]
        expired:
          type: ["boolean", "null"]
        signatureValid:
          type: ["boolean", "null"]
        issues:
          type: array
          items:
            type: string
    ReverseShellInput:
      type: object
      required: [lhost, lport]
      properties:
        lhost:
          type: string
        lport:
          type: integer
        languages:
          type: array
          items:
            type: string
        encoding:
          type: ["string", "null"]
    ReverseShellResult:
      type: object
      required: [shells]
      properties:
        shells:
          type: array
          items:
            $ref: "#/components/schemas/ReverseShell"
    ReverseShell:
      type: object
      required: [language, command, description]
      properties:
        language:
          type: string
        command:
          type: string
        description:
          type: string
        encoded:
          type: ["string", "null"]
    IpIntelInput:
      type: object
      required: [ip]
      properties:
        ip:
          type: string
        sources:
          type: array
          items:
            type: string
    IpIntelResult:
      type: object
      required: [ip]
      properties:
        ip:
          type: string
        country:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        asn:
          type: ["string", "null"]
        org:
          type: ["string", "null"]
        isTor:
          type: boolean
        isVpn:
          type: boolean
        isKnownMalicious:
          type: boolean
        abuseScore:
          type: ["integer", "null"]
        vtMalicious:
          type: ["integer", "null"]
        reverseDns:
          type: ["string", "null"]
        sources:
          type: array
          items:
            $ref: "#/components/schemas/IntelSource"
    IntelSource:
      type: object
      required: [name, available, data]
      properties:
        name:
          type: string
        available:
          type: boolean
        data:
          type: object
          additionalProperties: true
        error:
          type: ["string", "null"]
    HibpInput:
      type: object
      required: [type, value]
      properties:
        type:
          type: string
          enum: [email, password]
        value:
          type: string
    HibpResult:
      type: object
      required: [found, count]
      properties:
        found:
          type: boolean
        count:
          type: integer
        breaches:
          type: array
          items:
            $ref: "#/components/schemas/Breach"
    Breach:
      type: object
      required: [name, breachDate]
      properties:
        name:
          type: string
        breachDate:
          type: string
        dataClasses:
          type: array
          items:
            type: string
        description:
          type: ["string", "null"]
        isVerified:
          type: boolean

    # ── Plugins ─────────────────────────────────────────────────────────────
...