# lib/api-zod/src/generated/types/identifyHashResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GodsApp — Professional Security Auditing & Research Suite API
 * OpenAPI spec version: 1.0.0
 */
import type { HashCandidate } from './hashCandidate';

export interface IdentifyHashResult {
  hash: string;
  candidates: HashCandidate[];
}